fix bugs

+++ onnxruntime/test/providers/qnn/qnn_test_utils.h
@@ -1291,13 +1291,13 @@ inline GetTestModelFn BuildOpTestCase(const std::string& node_name,
     std::vector<std::string> op_input_names;
     op_input_names.reserve(input_defs_1.size() + input_defs_2.size());
 
-    for (int i = 0; i < input_defs_1.size(); i++) {
+    for (size_t i = 0; i < input_defs_1.size(); i++) {
       const std::string tmp_name = "input_defs_1_" + std::to_string(i);
       MakeTestInput<InputType1>(builder, tmp_name, input_defs_1[i], input_allocator);
       op_input_names.push_back(tmp_name);
     }
 
-    for (int i = 0; i < input_defs_2.size(); i++) {
+    for (size_t i = 0; i < input_defs_2.size(); i++) {
       const std::string tmp_name = "input_defs_2_" + std::to_string(i);
       MakeTestInput<InputType2>(builder, tmp_name, input_defs_2[i], input_allocator);
       op_input_names.push_back(tmp_name);
@@ -1327,19 +1327,19 @@ inline GetTestModelFn BuildOpTestCase(const std::string& node_name,
     std::vector<std::string> op_input_names;
     op_input_names.reserve(input_defs_1.size() + input_defs_2.size() + input_defs_3.size());
 
-    for (int i = 0; i < input_defs_1.size(); i++) {
+    for (size_t i = 0; i < input_defs_1.size(); i++) {
       const std::string tmp_name = "input_defs_1_" + std::to_string(i);
       MakeTestInput<InputType1>(builder, tmp_name, input_defs_1[i], input_allocator);
       op_input_names.push_back(tmp_name);
     }
 
-    for (int i = 0; i < input_defs_2.size(); i++) {
+    for (size_t i = 0; i < input_defs_2.size(); i++) {
       const std::string tmp_name = "input_defs_2_" + std::to_string(i);
       MakeTestInput<InputType2>(builder, tmp_name, input_defs_2[i], input_allocator);
       op_input_names.push_back(tmp_name);
     }
 
-    for (int i = 0; i < input_defs_3.size(); i++) {
+    for (size_t i = 0; i < input_defs_3.size(); i++) {
       const std::string tmp_name = "input_defs_3_" + std::to_string(i);
       MakeTestInput<InputType1>(builder, tmp_name, input_defs_3[i], input_allocator);
       op_input_names.push_back(tmp_name);
